+++ hybrid_os/hybrid_os_darwin/CHANGELOG.md
@@ -1,3 +1,11 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Add `Darwin` interface class.
+* Add `DarwinVersion` class.
+* Add `DarwinPlatform` class.
+* Add `iOS` interface class.
+* Add `iOSPlatform` class.
+* Add `HybridOSiOSPlugin` class.
+* Add `macOS` interface class.
+* Add `macOSPlatform` class.
+* Add `HybridOSmacOSPlugin` class.